+++ Senai.Operadores.Logicos/Senai.Operadores.Logicos.Exercicio3/Program.cs
@@ -5,10 +5,12 @@ namespace Senai.Operadores.Logicos.Exercicio3
     class Program
     {
         static void Main(string[] args)
-        {
+        {   
+            //solicita a senha
             Console.WriteLine("Insira a senha:");
-            double Senha = double.Parse(Console.ReadLine());
+            int Senha = int.Parse(Console.ReadLine());
 
+            //verifica se a senha está correta
             if(Senha==1234){
                 Console.WriteLine("ACESSO PERMITIDO");
             }else{
