+++ Senai.Operadores.Logicos/Senai.Operadores.Logicos.Exercicio4/Program.cs
@@ -6,9 +6,11 @@ namespace Senai.Operadores.Logicos.Exercicio4
     {
         static void Main(string[] args)
         {
+            //solicita o número do maçãs compradas
             Console.WriteLine("Insira o número de maçãs compradas:");
-            double Macas = double.Parse(Console.ReadLine());
+            int Macas = int.Parse(Console.ReadLine());
             
+            //verifica o número de maçãs compradas e calcula o preço total
             if(Macas>=12){
                 double Preco = 0.25 * Macas;
                 Console.WriteLine($"Valor total da compra: {Preco.ToString("c")}");
